requirements.txt
fastapi==0.100.0
motor==3.1.1
python-decouple==3.8
jinja2==3.1.2
pytest==7.4.0
httpx==0.24.1